round2_backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.4.1+cpu

fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
python-dotenv==1.0.1

numpy==1.26.4
faiss-cpu==1.7.4
sentence-transformers==2.7.0
rank-bm25==0.2.2          


PyMuPDF==1.24.9           
Pillow==10.3.0
pytesseract==0.3.10       


scikit-learn==1.4.2
langdetect==1.0.9
jsonschema==4.23.0
orjson==3.10.6


google-generativeai>=0.7.2

openai>=1.40.0          


httpx>=0.27.0            
slowapi==0.1.9          
limits>=3.11.0           
redis>=5.0.0            
